moved Currently Selected component position and adjust layout for class search
