+++ project/bldr/src/components/CalendarEditor.tsx
@@ -205,7 +205,7 @@ const CalendarEditor = () => {
                                                 {cls.pinned && (
                                                   <Pin className="h-2.5 w-2.5 lg:h-3 lg:w-3 text-amber-600 shrink-0" />
                                                 )}
-                                                <span className="truncate">
+                                                <span className="truncate select-none">
                                                   {cls.dept} {cls.code} (
                                                   {cls.component})
                                                 </span>
